changes for SFK-195

+++ lib/domain/data_uimodels/feedback_ui_model.dart
@@ -5,7 +5,7 @@ class FeedbackUiModel {
   final String? feedbackAuthorPhoto;
   final DateTime? feedbackDateTime;
   final String feedbackText;
-  final double? feedbackRating;
+  final int? feedbackRating;
   final UserTileType? feedbackAuthorType;
   final bool empty;
